extract utils to global util directory
